fixed issues after pulling main
