+++ Data/DTOs/Candidate/CandidateAddTechNoteDTO.cs
@@ -4,13 +4,13 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Services.DTOs.Candidate
+namespace Data.DTOs.Candidate
 {
-    public class CandidateAssigneesDTO
+    public class CandidateAddTechNoteDTO
     {
-        public int Id { get; set; }
+        public string Note { get; set; }
+        public int Score { get; set; }
         public int TechId { get; set; }
-        public int RecruiterId { get; set; }
         public DateTime? LastUpdatedDate { get; set; }
         public int LastUpdatedBy { get; set; }
     }
